Updated the objects to use Float. Decomposed the objects

+++ Beamformer/Beamformer/ChannelDelayData.swift
@@ -5,11 +5,11 @@ import Foundation
 public struct ChannelDelayData
 {
     public var channelIdentifier: Int
-    public var delays: [Double]
+    public var delays: [Float]
 
     public init(channelIdentifier: Int, numberOfDelays: Int)
     {
         self.channelIdentifier = channelIdentifier
-        self.delays = [Double](count: numberOfDelays, repeatedValue: 0)
+        self.delays = [Float](count: numberOfDelays, repeatedValue: 0)
     }
 }
